moved checkerstatus enum
