+++ src/app/(user)/(orders)/your-orders-as-checker/for-reupload/page.tsx
@@ -1,7 +1,7 @@
 "use client"
 
 import useSession from "@/hooks/useSession"
-import { CheckerStatus } from "../../_lib/constants"
+import CheckerStatus from "@/enums/checkerStatus"
 import YourOrdersPage from "../../_components/YourOrdersPage"
 
 export default function YourOrdersAsCheckerForReuploadPage() {
