+++ src/enums/checkerStatus.ts
@@ -1,0 +1,8 @@
+enum CheckerStatus {
+  Claimed = "Claimed",
+  ForReview = "For Review",
+  ReuploadRequested = "Reupload Requested",
+  Approved = "Approved",
+}
+
+export default CheckerStatus
